+++ src/Couchbase.EntityFrameworkCore/Query/Internal/Translators/CouchbaseStringMemberTranslator.cs
@@ -1,0 +1,58 @@
+using System.Reflection;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Diagnostics;
+using Microsoft.EntityFrameworkCore.Query;
+using Microsoft.EntityFrameworkCore.Query.SqlExpressions;
+
+namespace Couchbase.EntityFrameworkCore.Query.Internal.Translators;
+
+public class CouchbaseStringMemberTranslator : IMemberTranslator
+{
+    private static readonly MemberInfo LengthMemberInfo
+        = typeof(string).GetProperty(nameof(string.Length))!;
+
+    private readonly ISqlExpressionFactory _sqlExpressionFactory;
+
+    public CouchbaseStringMemberTranslator(ISqlExpressionFactory sqlExpressionFactory)
+    {
+        _sqlExpressionFactory = sqlExpressionFactory;
+    }
+
+    public virtual SqlExpression? Translate(
+        SqlExpression? instance,
+        MemberInfo member,
+        Type returnType,
+        IDiagnosticsLogger<DbLoggerCategory.Query> logger)
+    {
+        if (instance != null && LengthMemberInfo.Equals(member))
+        {
+            return _sqlExpressionFactory.Function(
+                "LENGTH",
+                new[] { instance },
+                nullable: true,
+                argumentsPropagateNullability: new[] { true },
+                returnType);
+        }
+
+        return null;
+    }
+}
+
+/* ************************************************************
+ *
+ *    @author Couchbase <info@couchbase.com>
+ *    @copyright 2025 Couchbase, Inc.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ *
+ * ************************************************************/